# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SaaS Platform API - Multi-tenant platform with auth, billing, and app registry
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /organizations/{orgId}/invitations:
    get:
      operationId: getOrgInvitations
      tags: [invitations]
      summary: Get pending invitations for an organization
      parameters:
        - in: path
          name: orgId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Invitations list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invitation"
    post:
      operationId: createInvitation
      tags: [invitations]
      summary: Invite a user to an organization
      parameters:
        - in: path
          name: orgId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvitationRequest"
      responses:
        "201":
          description: Invitation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invitation"
  /organizations/{orgId}/invitations/{invitationId}:
    delete:
      operationId: cancelInvitation
      tags: [invitations]
      summary: Cancel an invitation
      parameters:
        - in: path
          name: orgId
          required: true
          schema:
            type: string
        - in: path
          name: invitationId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Invitation cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /invitations/{token}/accept:
    post:
      operationId: acceptInvitation
      tags: [invitations]
      summary: Accept an invitation by token
      parameters:
        - in: path
          name: token
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Invitation accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"

  # ── APPS ──────────────────────────────────────────────────────────────────
...
  /billing/checkout:
    post:
      operationId: createCheckoutSession
      tags: [billing]
      summary: Create a Stripe Checkout session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutRequest"
      responses:
        "200":
          description: Checkout URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutResponse"
  /billing/portal:
    post:
      operationId: createPortalSession
      tags: [billing]
      summary: Create a Stripe Customer Portal session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PortalRequest"
      responses:
        "200":
          description: Portal URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortalResponse"
...
  /ayni/ceremonies:
    get:
      operationId: getAyniCeremonies
      tags: [ayni]
      summary: List ceremonies
      parameters:
        - in: query
          name: orgId
          schema:
            type: string
          required: true
        - in: query
          name: limit
          schema:
            type: integer
        - in: query
          name: offset
          schema:
            type: integer
      responses:
        "200":
          description: Ceremonies list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AyniCeremonyPage"
    post:
      operationId: createAyniCeremony
      tags: [ayni]
      summary: Create a ceremony
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAyniCeremonyRequest"
      responses:
        "201":
          description: Created ceremony
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AyniCeremony"
  /ayni/ceremonies/{ceremonyId}:
    get:
      operationId: getAyniCeremony
      tags: [ayni]
      summary: Get ceremony details
      parameters:
        - in: path
          name: ceremonyId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Ceremony details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AyniCeremony"
  /ayni/participants:
    get:
      operationId: getAyniParticipants
      tags: [ayni]
      summary: List participants for a ceremony
      parameters:
        - in: query
          name: ceremonyId
          schema:
            type: string
          required: true
      responses:
        "200":
          description: Participants list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AyniParticipant"
  /ayni/staff:
    get:
      operationId: getAyniStaff
      tags: [ayni]
      summary: List staff members for an organization
      parameters:
        - in: query
          name: orgId
          schema:
            type: string
          required: true
      responses:
        "200":
          description: Staff list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AyniStaff"
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    Organization:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        logoUrl:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        memberCount:
          type: integer
        stripeCustomerId:
          type: ["string", "null"]
      required: [id, name, slug, createdAt]
...
    Invitation:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [owner, admin, member, viewer]
        orgId:
          type: string
        orgName:
          type: string
        status:
          type: string
          enum: [pending, accepted, cancelled, expired]
        expiresAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required: [id, email, role, orgId, orgName, status, expiresAt, createdAt]
    CreateInvitationRequest:
      type: object
      properties:
        email:
          type: string
        role:
          type: string
          enum: [owner, admin, member, viewer]
      required: [email, role]

    # ── APPS ──────────────────────────────────────────────────────────────────
...
    CheckoutRequest:
      type: object
      properties:
        orgId:
          type: string
        appId:
          type: string
        planId:
          type: string
        successUrl:
          type: string
        cancelUrl:
          type: string
      required: [orgId, appId, planId]
    CheckoutResponse:
      type: object
      properties:
        url:
          type: string
      required: [url]
    PortalRequest:
      type: object
      properties:
        orgId:
          type: string
        returnUrl:
          type: string
      required: [orgId]
    PortalResponse:
      type: object
      properties:
        url:
          type: string
      required: [url]

    # ── AUDIT LOGS ────────────────────────────────────────────────────────────
...
    AyniCeremony:
      type: object
      properties:
        id:
          type: string
        orgId:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        scheduledAt:
          type: ["string", "null"]
          format: date-time
        location:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, scheduled, in_progress, completed, cancelled]
        capacity:
          type: ["integer", "null"]
        participantCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, orgId, name, status, participantCount, createdAt]
    CreateAyniCeremonyRequest:
      type: object
      properties:
        orgId:
          type: string
        name:
          type: string
        description:
          type: string
        scheduledAt:
          type: string
          format: date-time
        location:
          type: string
        capacity:
          type: integer
      required: [orgId, name]
    AyniCeremonyPage:
      type: object
      properties:
        ceremonies:
          type: array
          items:
            $ref: "#/components/schemas/AyniCeremony"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
      required: [ceremonies, total, limit, offset]
    AyniParticipant:
      type: object
      properties:
        id:
          type: string
        ceremonyId:
          type: string
        name:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        status:
          type: string
          enum: [registered, screened, confirmed, attended, cancelled]
        notes:
          type: ["string", "null"]
        registeredAt:
          type: string
          format: date-time
      required: [id, ceremonyId, name, status, registeredAt]
    AyniStaff:
      type: object
      properties:
        id:
          type: string
        orgId:
          type: string
        userId:
          type: ["string", "null"]
        name:
          type: string
        email:
          type: ["string", "null"]
        role:
          type: string
        isActive:
          type: boolean
      required: [id, orgId, name, role, isActive]